perf+test: keep the UI responsive under a bad daemon; harden the suite (R2, R11)
R2 - client responsiveness (src/pty.rs, src/runtime.rs):
- reattachment is now a queue drained under a 100 ms per-frame budget
  with exponential backoff. Previously every terminal was re-attached
  serially at 2 s each from the render thread, so a dead daemon with N
  terminals froze the UI for ~2N seconds on every retry frame.
- connect, autospawn-wait and hello moved to a connector thread whose
  result is only adopted on the render thread, so the owner's
  synchronous request/response model with idempotency keys is
  preserved. ensure_connected now fails visibly instead of blocking.
- sockets are shutdown() before being dropped, so the reader thread
  actually exits; previously each reconnect leaked a thread and an fd.
- drain_events budgets 128 messages / 256 KiB per frame.
- draw/poll/read errors are classified: transient kinds retry, fatal
  ones force a save checkpoint before returning. Raw-mode restore was
  already guaranteed by TerminalGuard; unsaved state was the real loss.

R11 - tests:
- CI now proves the PTY integration tests actually ran, via
  MULT_REQUIRE_PTY_INTEGRATION plus a sentinel emitted only after a real
  daemon drove a real PTY. The nix job legitimately skips them, and
  nothing previously asserted any job did not.
- fixed the scrollback test's race properly rather than widening its
  timeout: it awaited a live PtyOutput that, in the losing interleaving,
  never existed because the command finished before Attach. The pane now
  blocks on a FIFO whose open is the handshake, so the second write is
  live by construction. 15 consecutive runs green.
- new protocol framing suite: truncation at every prefix offset,
  zero-length frames, one-byte-per-read reassembly, EINTR retry,
  back-to-back frames, and 256 seeded postcard mutations.
- replay chunking tested at CHUNK-1/CHUNK/CHUNK+1/3*CHUNK+7.
- storage shape-error and backup-rename-failure behaviour pinned.
- env-mutating tests replaced with pure seams; no test mutates
  process-global env any more.
- agent.rs gained spawn-failure, nonzero-exit, invalid-UTF-8 and
  backpressure coverage, and states in-module that it is unwired
  Phase 3.4 scaffolding.
- integration deadlines raised and sleeps replaced with handshakes; the
  suite still finishes in under a second.

R11 confirms E11's premise: routing every serde error into
backup-and-reset costs the user every workspace for one renamed field,
with no notice. R7 changes that policy.

369 tests. Note: the previous commit (0ce4b08) accidentally swept an
in-progress test file from this slice into itself via git add -A while
an agent was still working; nothing was lost, but its contents are
broader than its message describes.

+++ crates/protocol/tests/framing.rs
@@ -72,9 +72,9 @@ fn identity() -> SessionIdentity {
     }
 }
 
-fn sample_message(session: u32) -> ClientMessage {
+fn sample_message(session: u64) -> ClientMessage {
     ClientMessage::Attach {
-        request_id: mult_protocol::RequestId::new(u64::from(session) + 1).expect("non-zero"),
+        request_id: mult_protocol::RequestId::new(session + 1).expect("non-zero"),
         identity: identity(),
         session: SessionId(session),
         rows: 24,
